requirements.txt
fastapi
uvicorn
moviepy
imageio
imageio-ffmpeg
numpy
pillow
pydantic
python-multipart
tqdm
gunicorn
google-genai
beautifulsoup4
requests
piper-tts
python-dotenv
groq
reportlab
